requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_libavif_api=1.0.4=h57928b3_4
aom=3.9.1=he0c23c2_0
asttokens=2.4.1=pyhd8ed1ab_0
blas=2.112=mkl
blas-devel=3.9.0=12_win64_mkl
brotli=1.1.0=hcfcfb64_1
brotli-bin=1.1.0=hcfcfb64_1
brotli-python=1.1.0=py39h99910a6_1
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2024.6.2=h56e8100_0
cairo=1.18.0=h1fef639_0
certifi=2024.6.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.2.1=py39h1f6ef14_0
cuda-cccl=12.5.39=0
cuda-cccl_win-64=12.5.39=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-opencl-dev=12.5.39=0
cuda-profiler-api=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cycler=0.12.1=pyhd8ed1ab_0
dav1d=1.2.1=hcfcfb64_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h63175ca_0
filelock=3.15.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_2
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.53.0=py39ha55e580_0
freetype=2.12.1=hdaf720e_2
glib=2.80.2=h7025463_1
glib-tools=2.80.2=h4394cf3_1
graphite2=1.3.13=h63175ca_1003
gst-plugins-base=1.24.4=hba88be7_0
gstreamer=1.24.4=h5006eae_0
harfbuzz=8.5.0=h81778c3_0
icu=73.2=h63175ca_0
idna=3.7=pyhd8ed1ab_0
imageio=2.34.1=pyh4b66e23_0
importlib-resources=6.4.0=pyhd8ed1ab_0
importlib_resources=6.4.0=pyhd8ed1ab_0
intel-openmp=2024.1.0=h57928b3_966
ipython=8.18.1=pyh7428d3b_3
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
kiwisolver=1.4.5=py39h1f6ef14_1
krb5=1.21.2=heb0366b_0
lcms2=2.16=h67d730c_0
lerc=4.0.0=h63175ca_0
libavif16=1.0.4=hac688ae_4
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.1.0=hcfcfb64_1
libbrotlidec=1.1.0=hcfcfb64_1
libbrotlienc=1.1.0=hcfcfb64_1
libcblas=3.9.0=12_win64_mkl
libclang13=18.1.7=default_h97ce8ae_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.6.39=0
libcurand-dev=10.3.6.39=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.20=hcfcfb64_0
libexpat=2.6.2=h63175ca_0
libffi=3.4.2=h8ffe710_5
libflac=1.4.3=h63175ca_0
libglib=2.80.2=h7025463_1
libhwloc=2.10.0=default_h8125262_1001
libiconv=1.17=hcfcfb64_2
libintl=0.22.5=h5728263_2
libintl-devel=0.22.5=h5728263_2
libjpeg-turbo=3.0.0=hcfcfb64_1
liblapack=3.9.0=12_win64_mkl
liblapacke=3.9.0=12_win64_mkl
libmad=0.15.1b=hcfcfb64_1001
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libogg=1.3.4=h8ffe710_1
libpng=1.6.43=h19919ed_0
libsqlite=3.46.0=h2466b09_0
libtiff=4.6.0=hddb2be6_3
libuv=1.48.0=hcfcfb64_0
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.4.0=hcfcfb64_0
libxcb=1.15=hcd874cb_0
libxml2=2.12.7=h283a6d9_1
libzlib=1.3.1=h2466b09_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.5=py39ha55989b_0
matplotlib=3.8.4=py39hcbf5309_2
matplotlib-base=3.8.4=py39he1095e7_2
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2021.4.0=h0e2418a_729
mkl-devel=2021.4.0=h57928b3_730
mkl-include=2021.4.0=h0e2418a_729
mpg123=1.32.6=h63175ca_0
mpmath=1.3.0=pyhd8ed1ab_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
networkx=3.2.1=pyhd8ed1ab_0
numpy=2.0.0=py39h60232e0_0
openjpeg=2.5.2=h3d672ee_0
openssl=3.3.1=h2466b09_0
packaging=24.1=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.44=h3d7b363_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py39h9ee4981_0
pip=24.0=pyhd8ed1ab_0
pixman=0.43.4=h63175ca_0
ply=3.11=pyhd8ed1ab_2
portmidi=2.0.4=h63175ca_2
prompt-toolkit=3.0.47=pyha770c72_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pure_eval=0.2.2=pyhd8ed1ab_0
pygame=2.5.2=py39h3d3616c_2
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pyhd8ed1ab_0
pyqt=5.15.9=py39hb77abff_5
pyqt5-sip=12.12.2=py39h99910a6_5
pysocks=1.7.1=pyh0701188_6
python=3.9.19=h4de0772_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.9=4_cp39
pytorch=2.3.1=py3.9_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py39ha55989b_1
qt-main=5.15.8=hcef0176_21
rav1e=0.6.6=h975169c_2
requests=2.32.3=pyhd8ed1ab_0
sdl2=2.30.2=h63175ca_0
sdl2_image=2.8.2=h5ca362a_1
sdl2_mixer=2.6.3=he8f0768_1
sdl2_ttf=2.22.0=hd5f56a6_1
setuptools=70.0.0=pyhd8ed1ab_0
sip=6.7.12=py39h99910a6_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.46.0=h2466b09_0
stack_data=0.6.2=pyhd8ed1ab_0
svt-av1=2.1.0=he0c23c2_0
sympy=1.12.1=pyh04b8f61_3
tbb=2021.12.0=hc790b64_1
tk=8.6.13=h5226925_1
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=2.3.1=pypi_0
torchvision=0.18.1=pypi_0
tornado=6.4.1=py39ha55e580_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h0c530f3_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=15.1.0=py39ha55989b_0
urllib3=2.2.1=pyhd8ed1ab_0
vc=14.3=h8a93ad2_20
vc14_runtime=14.40.33810=ha82c5b3_20
vs2015_runtime=14.40.33810=h3bf8584_20
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_1
win_inet_pton=1.1.0=pyhd8ed1ab_6
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.3.1=h2466b09_1
zstd=1.5.6=h0ea2cb4_0
